requirements.txt
Flask == 3.1.0
pybase64 == 1.4.1
google-cloud-vision == 3.10.1